Preprocessing and RandomForestClassifer model
Uploaded the initial dataset, looked through the set at the null values and preprocessed them with the preprocessing file. This file solves the issues, but gives a few pointers while running it on how to word it better, will look at this later. 
At the end of the preprocessing file it saves a new CSV file called RainAUS and this one is then used for the RandomForest model file.
RandomForest file works smoothly, uses Random Seed of 42.