Train and test split function made
Successful implementation of this function, to MASSIVELY reduce copying and pasting of code for each model code. Implemented in Knn and RandomForest files, need to change for knn plot code.
General idea of the function, input is the dataset and n is the percentage for splitting the dataset and rest is like we had before :)